ldd-check updates (#20)
* ldd-check - Copy pipeline from wolfi-os

This way we can maintain changes in "upstream" and then sync
them to wolfi when we release.

* ldd-check - update pipeline to use context.name by default for packages.

Melange v0.21.2 and newer have support for 'context.name' substitution
which means most tests can now just have:

    uses: test/tw/ldd-check

If you want to use 'files' and do _not_ want to use the default
packages value, then

    uses: test/tw/ldd-check
    with:
      packages: none
      files: /lib/my.library.so

* ldd-check - Add support for '--packages=none' for explicit empty packages.

* ldd-check - Allow both --packages and --files to be provided.

This actually works in a reasonable way.  The files get
counted but are required to exist and be something that
ldd can check, but contents of '--packages' are allowed to be
"not a dynamic executable".

Also move 'passes' and 'fails' toward the bottom, closer to the
"main".