feat(ldd-check): allow to exclude binaries to analyze
This commit adds a new flag --exclude-files to allow to skip binary
files to run ldd on.
This is useful in cases where we dont know in advance how many binaries
and which binaries are packaged that we want to ensure shared objects
they depend on exist, but we know in advance which ones we want to ignore.

Signed-off-by: Massimiliano Giovagnoli <massimiliano.giovagnoli@chainguard.dev>