Update setup-melange action to v1.4.6
A change in the melange action means that we do not
need to do the chown of ~/go/pkg/mod any more.
In fact attempting to do so results in

    ls: cannot access '/home/runner/go/pkg/mod': No such file or directory

I'm not exactly sure why the change in behavior from melange,
but the code was only here to workaround behavior that *was* there.